+++ relations/src/r1cs/constraint_system.rs
@@ -767,6 +767,16 @@ impl<F: Field> ConstraintSystemRef<F> {
         }
     }
 
+    /// Consumes self to return the inner `ConstraintSystem<F>`. Returns
+    /// `None` if `Self::CS` is `None` or if any other references to
+    /// `Self::CS` exist.  
+    pub fn into_inner(self) -> Option<ConstraintSystem<F>> {
+        match self {
+            Self::CS(a) => Rc::try_unwrap(a).ok().map(|s| s.into_inner()),
+            Self::None => None,
+        }
+    }
+
     /// Obtain an immutable reference to the underlying `ConstraintSystem`.
     ///
     /// # Panics